% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{out_hampel_filter}
\alias{out_hampel_filter}
\title{Hampel filter}
\usage{
out_hampel_filter(y, k = 5L, t0 = 3L, method = "hampel",
  parent_logger = "test")
}
\arguments{
\item{y}{vector of values for outlier substitution}

\item{k}{window semi-value (integer) for substitution value calculation. See
details.}

\item{t0}{}

\item{method}{Character vector indicating the method to use in the estimation:
\code{hampel} (default), \code{tukey} and \code{quantile}.}
}
\value{
A list with two elements:
  \itemize{
    \item{res: A vector of the same length as y with the ouliers value
    substituted by the estimation}
    \item{index: A vector indicating which TIMESTAMPS have been modified}
  }
}
\description{
Hampel filter for detecting and substituting outliers in the environmental
and sap flow data.
}
\details{
This is a modified version of the Hampel filter to avoid breaking when NAs
are present in the data. Options include the posibility of using the
median (classic Hampel, default), tukeyline or quantile regression estimations.
}
\seealso{
Other outliers: \code{\link{out_remove}},
  \code{\link{outlier_subs}}
}

