% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{out_remove}
\alias{out_remove}
\title{Outliers QC}
\usage{
out_remove(sfn_data, k = 5L, t0 = 3L, method = "hampel",
  parent_logger = "test")
}
\arguments{
\item{sfn_data}{SfnData object with the site data and metadata}

\item{k}{window semi-value (integer) for substitution value calculation. See
\code{\link{out_hampel_filter}} for details.}

\item{t0}{}

\item{method}{Character vector indicating the method to use in the outlier
estimation: \code{hampel} (default), \code{tukey} and \code{quantile}.}
}
\value{
a SfnData object as the one provided with the oulier values
  substituted and the flags slot updated.
}
\description{
Outliers detection, substitution and annotation for SfnData objects
}
\details{
Outliers for sap flow data and environmental data are located and substituted
by the selected method, using the \code{\link{out_hampel_filter}} function.
}
\seealso{
Other outliers: \code{\link{out_hampel_filter}},
  \code{\link{outlier_subs}}
}

